% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_ibp_cartesian.R
\name{coord_ibp_flip}
\alias{coord_ibp_flip}
\title{Title}
\usage{
coord_ibp_flip(ratio = 1, xlim = NULL, ylim = NULL, expand = TRUE, clip = "on")
}
\arguments{
\item{clip}{}
}
\value{

}
\description{
Title
}
