% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_breaks.R
\name{equal_breaks}
\alias{equal_breaks}
\title{Title}
\usage{
equal_breaks(
  n = 3,
  s = 0.05,
  manual = NULL,
  remove_border = FALSE,
  leading = 1,
  ...
)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Title
}
