% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_fraunhofer_label_datetime.R
\name{geom_fraunhofer_label_datetime}
\alias{geom_fraunhofer_label_datetime}
\title{Title}
\usage{
geom_fraunhofer_label_datetime(
  ...,
  angle = 90,
  hjust = -0.05,
  vjust = -0.5,
  x_pos = 1,
  y_pos = -1,
  face = "plain",
  font = "Frutiger 45 Light"
)
}
\arguments{
\item{angle}{Winkel in dem der Schriftzug angezeigt wird, Standard 90 Grad.}

\item{hjust}{Horizontale Ausrichtung vom Ankerpunkt. Wir genutzt BEVOR der Text gedreht wird.}

\item{vjust}{Vertikale Ausrichtung vom Ankerpunkt. Wir genutzt BEVOR der Text gedreht wird.}

\item{x_pos}{Setzt den Ankerpunkt bei min(x) (x_pos==-1) oder max(x) (x_pos==1)}

\item{y_pos}{Setzt den Ankerpunkt bei min(y) (y_pos==-1) oder max(y) (y_pos==1)}
}
\value{

}
\description{
Title
}
