% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ibp.R
\name{gg_ibp}
\alias{gg_ibp}
\title{Title}
\usage{
gg_ibp(
  n_x = 7,
  n_y = 5,
  limits_x = NULL,
  limits_y = NULL,
  font = "Frutiger 45 Light",
  x_pos_leg = 0.1,
  y_pos_leg = 0.85,
  leading_x = 2,
  leading_y = 2
)
}
\arguments{
\item{y_pos_leg}{}
}
\value{

}
\description{
Title
}
