% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ibp_datetime.R
\name{gg_ibp_datetime}
\alias{gg_ibp_datetime}
\title{Title}
\usage{
gg_ibp_datetime(
  n_x = 7,
  n_y = 5,
  limits_x = NULL,
  limits_y = NULL,
  font = "Frutiger 45 Light",
  leading_y = 2
)
}
\arguments{
\item{y_pos_leg}{}
}
\value{

}
\description{
Title
}
