% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp_cols.R
\docType{data}
\name{ipb_all}
\alias{ipb_all}
\title{Alle CI Farben}
\format{A prismatic vector}
\usage{
ibp_all
}
\description{
Alle Fraunhofer CI-Farben (nach Skript von Matthias Winkler)
}
\keyword{datasets}
