% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp_cols.R
\docType{data}
\name{ipb_blue}
\alias{ipb_blue}
\title{Blau Abstufungen von Faunhofer}
\format{
A prismatic vector
}
\usage{
ibp_blue
}
\description{
Enthält Blauabstufungen der Fraunhofer CI (nach Matthias Winkler)
}
\keyword{datasets}
