% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp_cols.R
\docType{data}
\name{ipb_cols}
\alias{ipb_cols}
\title{Alle Fraunhofer Farben}
\format{
A List with multpiple Prismatic Vectors
}
\usage{
ibp_cols
}
\description{
Color sets with the Fraunhofer CI; Enthählt auch weiß-Schwarz Abstufungen der Hauptfarben.
Alle Farben sind auch noch mal als einzelne Vektoren verfügbar ibp_***
}
\keyword{datasets}
