% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp_cols.R
\docType{data}
\name{ipb_green}
\alias{ipb_green}
\title{Grün Abstufungen von Fraunhofer}
\format{
A prismatic vector
}
\usage{
ibp_green
}
\description{
Enthält Grünabstufungen der Fraunhofer CI (nach Matthias Winkler)
}
\keyword{datasets}
