% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp_cols.R
\docType{data}
\name{ipb_grey}
\alias{ipb_grey}
\title{Grau Abstufungen von Fraunhofer}
\format{A prismatic vector}
\usage{
ibp_grey
}
\description{
Enthält Grauabstufungen der Fraunhofer CI (nach Matthias Winkler)
}
\keyword{datasets}
