% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp_cols.R
\docType{data}
\name{ipb_logo}
\alias{ipb_logo}
\title{Abstufungen der Fraunhofer Titelfarbe}
\format{A prismatic vector}
\usage{
ibp_logo
}
\description{
Enthält Abstufungen der Fraunhofer CI (nach Matthias Winkler)
}
\keyword{datasets}
