% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp_cols.R
\docType{data}
\name{ipb_old}
\alias{ipb_old}
\title{Alte CI von Faunhofer}
\format{
A prismatic vector
}
\usage{
ibp_old
}
\description{
Enthält die alte Fraunhofer CI (nach Matthias Winkler)
}
\keyword{datasets}
