% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp_cols.R
\docType{data}
\name{ipb_old_matt}
\alias{ipb_old_matt}
\title{Alte CI von Faunhofer (matt)}
\format{
A prismatic vector
}
\usage{
ibp_old_matt
}
\description{
Enthält die alte Fraunhofer CI (matt) (nach Matthias Winkler)
}
\keyword{datasets}
