% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp_cols.R
\docType{data}
\name{ipb_oran}
\alias{ipb_oran}
\title{Orange Abstufungen von Fraunhofer}
\format{A prismatic vector}
\usage{
ibp_orang
}
\description{
Enthält orangeabstufungen der Fraunhofer CI (nach Matthias Winkler)
}
\keyword{datasets}
