% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp_cols.R
\docType{data}
\name{ipb_ppoint}
\alias{ipb_ppoint}
\title{Alle CI Farben nach der aktuellen PowerPoint Vorlage}
\format{
A prismatic vector
}
\usage{
ibp_ppoint
}
\description{
Alle Fraunhofer CI-Farben nach der aktuellen PowerPoint Vorlage
}
\keyword{datasets}
