% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp_cols.R
\docType{data}
\name{ipb_tuer}
\alias{ipb_tuer}
\title{Tuerkis Abstufungen von Fraunhofer}
\format{A prismatic vector}
\usage{
ibp_tuer
}
\description{
Enthält tuerkisabstufungen der Fraunhofer CI (nach Matthias Winkler)
}
\keyword{datasets}
