% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_last_label_dates.R
\name{remove_last_label_dates}
\alias{remove_last_label_dates}
\title{Title}
\usage{
remove_last_label_dates(formatation = "")
}
\arguments{
\item{formatation}{}
}
\value{

}
\description{
Title
}
