% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_ibp_datetime.R,
%   R/scale_y_ibp_datetime.R
\name{scale_x_ibp_datetime}
\alias{scale_x_ibp_datetime}
\title{Title}
\usage{
scale_x_ibp_datetime(n = 7, limits = NULL, remove_border_val = FALSE, ...)

scale_x_ibp_datetime(n = 7, limits = NULL, remove_border_val = FALSE, ...)
}
\arguments{
\item{n}{Anzahl von Major Gridlines}

\item{limits}{Vector mir minimal und maximal Wert}

\item{remove_border_val}{Derzeit nicht benutzt}

\item{...}{}
}
\value{



}
\description{
Title

Title
}
