% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_ibp_cont.R
\name{scale_y_ibp_cont}
\alias{scale_y_ibp_cont}
\title{Title}
\usage{
scale_y_ibp_cont(n = 5, limits = NULL, remove_border_val = FALSE, ...)
}
\arguments{
\item{n}{Anzahl von Major Gridlines}

\item{limits}{Vector mir minimal und maximal Wert}

\item{remove_border_val}{Derzeit nicht benutzt}

\item{...}{}
}
\value{

}
\description{
Title
}
