% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_less_labels_dates.R
\name{show_less_labels_dates}
\alias{show_less_labels_dates}
\title{Title}
\usage{
show_less_labels_dates(every_n = 2, formatation = "")
}
\arguments{
\item{formatation}{}
}
\value{

}
\description{
Title
}
