% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ibp.R
\name{theme_ibp}
\alias{theme_ibp}
\title{Title}
\usage{
theme_ibp(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/170,
  base_rect_size = base_size/170,
  x_pos_leg = 0.1,
  y_pos_leg = 0.85,
  face = "plain",
  ibp_family = "Frutiger 45 Light"
)
}
\arguments{
\item{base_size}{Schriftgroesse. Geklaut aus online Vorlage.}

\item{base_family}{Schriftart. "" ist sans-serif. Geklaut aus online Vorlage.}

\item{base_line_size}{Geklaut aus online Vorlage.}

\item{base_rect_size}{Geklaut aus online Vorlage.}

\item{x_pos_leg}{X-position der Legende. Standard links oben, Werte von 0 (min(x)) bis 1 (max(x))}

\item{y_pos_leg}{Y-position der Legende. Standard links oben, Werte von 0 (min(y)) bis 1 (max(y))}

\item{aspect}{Aspect ratio. Sollte ungefähr ny/nx sein}

\item{font}{Schriftart. Standard ist Frutiger 45 Light, durch die Bibliothek extrafont!}
}
\value{

}
\description{
Title
}
